@@ -9,7 +9,7 @@
   },
   "dependencies": {
     "expo": "~36.0.2",
-    "react": "~16.9.0",
+    "react": "~16.12.0",
     "react-dom": "~16.9.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-36.0.0.tar.gz",
     "react-native-autogrow-input": "^0.2.1",
